post-processing-scripts/requirements.txt
beautifulsoup4==4.11.2
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.0.1
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
python-decouple==3.7
requests==2.28.2
rfc3986==1.5.0
sniffio==1.3.0
soupsieve==2.4
urllib3==1.26.14
